% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{Calculates Trade Values}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
Calculates Trade Values
}
\keyword{datasets}
