% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values_generate}
\alias{values_generate}
\title{Calculates Trade Values}
\usage{
values_generate(
  players_raw,
  picks_raw,
  qb_type = "1QB",
  league_size = "12 teams",
  value_factor = 235,
  rookie_optimism = 75,
  draft_type = "Normal",
  future_factor = 80
)
}
\description{
Calculates Trade Values
}
